+++ sc-kpm/sc-common/sc_keynodes.h
@@ -12,11 +12,14 @@
 
 #include "sc_common_types.h"
 
-_SC_KPM_EXTERN sc_result
-sc_common_resolve_keynode(sc_memory_context const * ctx, char const * sys_idtf, sc_addr * keynode);
+_SC_KPM_EXTERN sc_result sc_common_resolve_keynode(
+    sc_memory_context * ctx,
+    char const * sys_idtf,
+    sc_addr * keynode,
+    sc_addr const init_memory_generated_structure);
 
-#define RESOLVE_KEYNODE(ctx, keynode) \
-  if (sc_common_resolve_keynode(ctx, keynode##_str, &keynode) != SC_RESULT_OK) \
+#define RESOLVE_KEYNODE(ctx, keynode, init_memory_generated_structure) \
+  if (sc_common_resolve_keynode(ctx, keynode##_str, &keynode, init_memory_generated_structure) != SC_RESULT_OK) \
   return SC_RESULT_ERROR
 
 #endif