+++ sc-kpm/sc-search/search.c
@@ -27,11 +27,11 @@ sc_event * event_question_search_links_of_relation_connected_with_element;
 
 // --------------------- Module ------------------------
 
-sc_result sc_module_initialize()
+sc_result sc_module_initialize_with_init_memory_generated_structure(sc_addr const init_memory_generated_structure)
 {
   s_default_ctx = sc_memory_context_new(sc_access_lvl_make_min);
 
-  if (search_keynodes_initialize(s_default_ctx) != SC_RESULT_OK)
+  if (search_keynodes_initialize(s_default_ctx, init_memory_generated_structure) != SC_RESULT_OK)
     return SC_RESULT_ERROR;
 
   event_question_search_all_output_arcs = sc_event_new(
@@ -120,6 +120,11 @@ sc_result sc_module_initialize()
     return SC_RESULT_ERROR;
 
   return SC_RESULT_OK;
+}
+
+_SC_EXT_EXTERN sc_uint32 sc_module_load_priority()
+{
+  return 0;
 }
 
 sc_result sc_module_shutdown()