+++ sc-kpm/sc-ui/ui.h
@@ -15,7 +15,8 @@ extern "C"
 extern "C"
 {
   // --------------- Module ----------------
-  _SC_EXT_EXTERN sc_result sc_module_initialize();
+  _SC_EXT_EXTERN sc_result
+  sc_module_initialize_with_init_memory_generated_structure(sc_addr const init_memory_generated_structure);
   _SC_EXT_EXTERN sc_result sc_module_shutdown();
 }