+++ sc-kpm/sc-utils/utils.c
@@ -6,24 +6,29 @@
 
 #include "utils.h"
 #include "utils_keynodes.h"
-#include "utils_garbage_deletion.h"
+#include "utils_erase_elements.h"
+#include "../sc-search/search_keynodes.h"
 
-sc_memory_context * s_default_ctx = 0;
-sc_memory_context * s_garbage_ctx = 0;
+sc_memory_context * s_erase_elements_ctx = 0;
 
-sc_event * event_garbage_deletion;
+sc_event * event_erase_elements;
 
-_SC_EXT_EXTERN sc_result sc_module_initialize()
+_SC_EXT_EXTERN sc_result
+sc_module_initialize_with_init_memory_generated_structure(sc_addr const init_memory_generated_structure)
 {
-  s_default_ctx = sc_memory_context_new(sc_access_lvl_make_min);
-  s_garbage_ctx = sc_memory_context_new(sc_access_lvl_make_max);
+  s_erase_elements_ctx = sc_memory_context_new(sc_access_lvl_make_max);
 
-  if (utils_keynodes_initialize() != SC_RESULT_OK)
+  if (utils_keynodes_initialize(init_memory_generated_structure) != SC_RESULT_OK)
     return SC_RESULT_ERROR;
 
-  event_garbage_deletion =
-      sc_event_new(s_default_ctx, keynode_sc_garbage, SC_EVENT_ADD_OUTPUT_ARC, 0, agent_garbage_delete, 0);
-  if (event_garbage_deletion == null_ptr)
+  event_erase_elements = sc_event_new(
+      s_erase_elements_ctx,
+      keynode_question_initiated,
+      SC_EVENT_ADD_OUTPUT_ARC,
+      null_ptr,
+      agent_erase_elements,
+      null_ptr);
+  if (event_erase_elements == null_ptr)
     return SC_RESULT_ERROR;
 
   return SC_RESULT_OK;
@@ -31,18 +36,17 @@ _SC_EXT_EXTERN sc_result sc_module_initialize()
 
 _SC_EXT_EXTERN sc_uint32 sc_module_load_priority()
 {
-  return 0;
+  return 1000;
 }
 
 _SC_EXT_EXTERN sc_result sc_module_shutdown()
 {
   sc_result res = SC_RESULT_OK;
 
-  if (event_garbage_deletion)
-    sc_event_destroy(event_garbage_deletion);
+  if (event_erase_elements)
+    sc_event_destroy(event_erase_elements);
 
-  sc_memory_context_free(s_garbage_ctx);
-  sc_memory_context_free(s_default_ctx);
+  sc_memory_context_free(s_erase_elements_ctx);
 
   return res;
 }