+++ sc-kpm/sc-utils/utils.h
@@ -9,10 +9,10 @@
 
 #include "sc-core/sc_memory.h"
 
-extern sc_memory_context * s_default_ctx;
-extern sc_memory_context * s_garbage_ctx;
+extern sc_memory_context * s_erase_elements_ctx;
 
-_SC_EXT_EXTERN sc_result sc_module_initialize();
+_SC_EXT_EXTERN sc_result
+sc_module_initialize_with_init_memory_generated_structure(sc_addr const init_memory_generated_structure);
 _SC_EXT_EXTERN sc_uint32 sc_module_load_priority();
 _SC_EXT_EXTERN sc_result sc_module_shutdown();