+++ sc-kpm/tests/sc-agents/sc-core-agents.cpp
@@ -7,13 +7,16 @@ extern "C"
 #include "sc-search/search.h"
 #include "sc-search/search_keynodes.h"
 #include "sc-search/search_agents.h"
+#include "sc-utils/utils_keynodes.h"
 }
 
 TEST_F(ScMemoryTest, agent_search_all_const_pos_output_arc)
 {
   sc_memory_context * context = sc_memory_context_new(sc_access_lvl_make_min);
+  sc_addr init_memory_generated_structure;
+  SC_ADDR_MAKE_EMPTY(init_memory_generated_structure);
 
-  sc_module_initialize();
+  sc_module_initialize_with_init_memory_generated_structure(init_memory_generated_structure);
 
   sc_addr const question = sc_memory_node_new(context, sc_type_node | sc_type_const);
   sc_addr const setAddr = sc_memory_node_new(context, sc_type_node | sc_type_const);
@@ -42,11 +45,7 @@ TEST_F(ScMemoryTest, agent_search_all_const_pos_output_arc)
   EXPECT_TRUE(sc_iterator5_next(it5));
   sc_iterator5_free(it5);
 
-  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(
-      context,
-      sc_iterator5_value(it5, 2),
-      sc_type_arc_pos_const_perm,
-      0);
+  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(context, sc_iterator5_value(it5, 2), sc_type_arc_pos_const_perm, 0);
 
   EXPECT_TRUE(sc_iterator3_next(it3));
   EXPECT_TRUE(sc_iterator3_next(it3));
@@ -63,7 +62,10 @@ TEST_F(ScMemoryTest, agent_search_all_const_pos_output_arc_with_rel)
 {
   sc_memory_context * context = sc_memory_context_new(sc_access_lvl_make_min);
 
-  sc_module_initialize();
+  sc_addr init_memory_generated_structure;
+  SC_ADDR_MAKE_EMPTY(init_memory_generated_structure);
+
+  sc_module_initialize_with_init_memory_generated_structure(init_memory_generated_structure);
 
   sc_addr const question = sc_memory_node_new(context, sc_type_node | sc_type_const);
   sc_addr const setAddr = sc_memory_node_new(context, sc_type_node | sc_type_const);
@@ -94,11 +96,7 @@ TEST_F(ScMemoryTest, agent_search_all_const_pos_output_arc_with_rel)
   EXPECT_TRUE(sc_iterator5_next(it5));
   sc_iterator5_free(it5);
 
-  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(
-      context,
-      sc_iterator5_value(it5, 2),
-      sc_type_arc_pos_const_perm,
-      0);
+  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(context, sc_iterator5_value(it5, 2), sc_type_arc_pos_const_perm, 0);
 
   EXPECT_TRUE(sc_iterator3_next(it3));
   EXPECT_TRUE(sc_iterator3_next(it3));
@@ -120,7 +118,10 @@ TEST_F(ScMemoryTest, agent_search_all_const_pos_input_arc)
 {
   sc_memory_context * context = sc_memory_context_new(sc_access_lvl_make_min);
 
-  sc_module_initialize();
+  sc_addr init_memory_generated_structure;
+  SC_ADDR_MAKE_EMPTY(init_memory_generated_structure);
+
+  sc_module_initialize_with_init_memory_generated_structure(init_memory_generated_structure);
 
   sc_addr const question = sc_memory_node_new(context, sc_type_node | sc_type_const);
   sc_addr const setAddr = sc_memory_node_new(context, sc_type_node | sc_type_const);
@@ -149,11 +150,7 @@ TEST_F(ScMemoryTest, agent_search_all_const_pos_input_arc)
   EXPECT_TRUE(sc_iterator5_next(it5));
   sc_iterator5_free(it5);
 
-  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(
-      context,
-      sc_iterator5_value(it5, 2),
-      sc_type_arc_pos_const_perm,
-      0);
+  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(context, sc_iterator5_value(it5, 2), sc_type_arc_pos_const_perm, 0);
 
   EXPECT_TRUE(sc_iterator3_next(it3));
   EXPECT_TRUE(sc_iterator3_next(it3));
@@ -173,7 +170,10 @@ TEST_F(ScMemoryTest, agent_search_all_const_pos_input_arc_with_rel)
 {
   sc_memory_context * context = sc_memory_context_new(sc_access_lvl_make_min);
 
-  sc_module_initialize();
+  sc_addr init_memory_generated_structure;
+  SC_ADDR_MAKE_EMPTY(init_memory_generated_structure);
+
+  sc_module_initialize_with_init_memory_generated_structure(init_memory_generated_structure);
 
   sc_addr const question = sc_memory_node_new(context, sc_type_node | sc_type_const);
   sc_addr const setAddr = sc_memory_node_new(context, sc_type_node | sc_type_const);
@@ -204,11 +204,7 @@ TEST_F(ScMemoryTest, agent_search_all_const_pos_input_arc_with_rel)
   EXPECT_TRUE(sc_iterator5_next(it5));
   sc_iterator5_free(it5);
 
-  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(
-      context,
-      sc_iterator5_value(it5, 2),
-      sc_type_arc_pos_const_perm,
-      0);
+  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(context, sc_iterator5_value(it5, 2), sc_type_arc_pos_const_perm, 0);
 
   EXPECT_TRUE(sc_iterator3_next(it3));
   EXPECT_TRUE(sc_iterator3_next(it3));
@@ -230,7 +226,10 @@ TEST_F(ScMemoryTest, agent_search_all_identifiers)
 {
   sc_memory_context * context = sc_memory_context_new(sc_access_lvl_make_min);
 
-  sc_module_initialize();
+  sc_addr init_memory_generated_structure;
+  SC_ADDR_MAKE_EMPTY(init_memory_generated_structure);
+
+  sc_module_initialize_with_init_memory_generated_structure(init_memory_generated_structure);
 
   sc_addr const question = sc_memory_node_new(context, sc_type_node | sc_type_const);
   sc_addr const setAddr = sc_memory_node_new(context, sc_type_node | sc_type_const);
@@ -264,11 +263,7 @@ TEST_F(ScMemoryTest, agent_search_all_identifiers)
   EXPECT_TRUE(sc_iterator5_next(it5));
   sc_iterator5_free(it5);
 
-  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(
-      context,
-      sc_iterator5_value(it5, 2),
-      sc_type_arc_pos_const_perm,
-      0);
+  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(context, sc_iterator5_value(it5, 2), sc_type_arc_pos_const_perm, 0);
 
   EXPECT_TRUE(sc_iterator3_next(it3));
   EXPECT_TRUE(sc_iterator3_next(it3));
@@ -283,7 +278,10 @@ TEST_F(ScMemoryTest, agent_search_all_identified_elements)
 {
   sc_memory_context * context = sc_memory_context_new(sc_access_lvl_make_min);
 
-  sc_module_initialize();
+  sc_addr init_memory_generated_structure;
+  SC_ADDR_MAKE_EMPTY(init_memory_generated_structure);
+
+  sc_module_initialize_with_init_memory_generated_structure(init_memory_generated_structure);
 
   sc_addr const setAddr = sc_memory_node_new(context, sc_type_node | sc_type_link);
 
@@ -313,11 +311,7 @@ TEST_F(ScMemoryTest, agent_search_all_identified_elements)
   EXPECT_TRUE(sc_iterator5_next(it5));
   sc_iterator5_free(it5);
 
-  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(
-      context,
-      sc_iterator5_value(it5, 2),
-      sc_type_arc_pos_const_perm,
-      0);
+  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(context, sc_iterator5_value(it5, 2), sc_type_arc_pos_const_perm, 0);
 
   EXPECT_TRUE(sc_iterator3_next(it3));
   EXPECT_TRUE(sc_iterator3_next(it3));
@@ -338,7 +332,10 @@ TEST_F(ScMemoryTest, agent_search_full_semantic_neighborhood)
 {
   sc_memory_context * context = sc_memory_context_new(sc_access_lvl_make_min);
 
-  sc_module_initialize();
+  sc_addr init_memory_generated_structure;
+  SC_ADDR_MAKE_EMPTY(init_memory_generated_structure);
+
+  sc_module_initialize_with_init_memory_generated_structure(init_memory_generated_structure);
 
   sc_addr const question = sc_memory_node_new(context, sc_type_node | sc_type_const);
   sc_addr const setAddr = sc_memory_node_new(context, sc_type_node | sc_type_const);
@@ -350,8 +347,8 @@ TEST_F(ScMemoryTest, agent_search_full_semantic_neighborhood)
   sc_memory_arc_new(context, sc_type_arc_pos_const_perm, keynode_nrel_main_idtf, idtfEdgeAddr);
 
   sc_addr const translationAddr = sc_memory_node_new(context, sc_type_node | sc_type_const);
-  sc_addr const translationEdgeAddr
-      = sc_memory_arc_new(context, sc_type_arc_common | sc_type_const, translationAddr, setAddr);
+  sc_addr const translationEdgeAddr =
+      sc_memory_arc_new(context, sc_type_arc_common | sc_type_const, translationAddr, setAddr);
   sc_memory_arc_new(context, sc_type_arc_pos_const_perm, keynode_nrel_translation, translationEdgeAddr);
 
   sc_addr const linkAddr = sc_memory_node_new(context, sc_type_link | sc_type_const);
@@ -363,8 +360,8 @@ TEST_F(ScMemoryTest, agent_search_full_semantic_neighborhood)
   sc_addr const noroleAddr = sc_memory_node_new(context, sc_type_node | sc_type_const | sc_type_node_norole);
 
   sc_addr const decompositionAddr = sc_memory_node_new(context, sc_type_node | sc_type_const);
-  sc_addr const decompositionEdge
-      = sc_memory_arc_new(context, sc_type_arc_common | sc_type_const, decompositionAddr, setAddr);
+  sc_addr const decompositionEdge =
+      sc_memory_arc_new(context, sc_type_arc_common | sc_type_const, decompositionAddr, setAddr);
   sc_memory_arc_new(context, sc_type_arc_pos_const_perm, noroleAddr, decompositionEdge);
   sc_memory_arc_new(context, sc_type_arc_pos_const_perm, keynode_quasybinary_relation, noroleAddr);
   sc_memory_arc_new(context, sc_type_arc_pos_const_perm, keynode_nonbinary_relation, noroleAddr);
@@ -392,11 +389,7 @@ TEST_F(ScMemoryTest, agent_search_full_semantic_neighborhood)
   EXPECT_TRUE(sc_iterator5_next(it5));
   sc_iterator5_free(it5);
 
-  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(
-      context,
-      sc_iterator5_value(it5, 2),
-      sc_type_arc_pos_const_perm,
-      0);
+  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(context, sc_iterator5_value(it5, 2), sc_type_arc_pos_const_perm, 0);
 
   EXPECT_TRUE(sc_iterator3_next(it3));
   EXPECT_TRUE(sc_iterator3_next(it3));
@@ -430,7 +423,10 @@ TEST_F(ScMemoryTest, agent_search_links_of_relation_connected_with_element)
 {
   sc_memory_context * context = sc_memory_context_new(sc_access_lvl_make_min);
 
-  sc_module_initialize();
+  sc_addr init_memory_generated_structure;
+  SC_ADDR_MAKE_EMPTY(init_memory_generated_structure);
+
+  sc_module_initialize_with_init_memory_generated_structure(init_memory_generated_structure);
 
   sc_addr const question = sc_memory_node_new(context, sc_type_node | sc_type_const);
   sc_addr const setAddr = sc_memory_node_new(context, sc_type_node | sc_type_const);
@@ -443,8 +439,8 @@ TEST_F(ScMemoryTest, agent_search_links_of_relation_connected_with_element)
   sc_memory_arc_new(context, sc_type_arc_pos_const_perm, keynode_nrel_main_idtf, idtfEdgeAddr);
 
   sc_addr const translationAddr = sc_memory_node_new(context, sc_type_node | sc_type_const);
-  sc_addr const translationEdgeAddr
-      = sc_memory_arc_new(context, sc_type_arc_common | sc_type_const, translationAddr, setAddr);
+  sc_addr const translationEdgeAddr =
+      sc_memory_arc_new(context, sc_type_arc_common | sc_type_const, translationAddr, setAddr);
   sc_memory_arc_new(context, sc_type_arc_pos_const_perm, keynode_nrel_translation, translationEdgeAddr);
 
   sc_addr const linkAddr = sc_memory_node_new(context, sc_type_link | sc_type_const);
@@ -458,8 +454,8 @@ TEST_F(ScMemoryTest, agent_search_links_of_relation_connected_with_element)
   sc_memory_arc_new(context, sc_type_arc_pos_const_perm, keynode_rrel_2, edge);
 
   sc_addr const decompositionAddr = sc_memory_node_new(context, sc_type_node | sc_type_const);
-  sc_addr const decompositionEdge
-      = sc_memory_arc_new(context, sc_type_arc_common | sc_type_const, decompositionAddr, setAddr);
+  sc_addr const decompositionEdge =
+      sc_memory_arc_new(context, sc_type_arc_common | sc_type_const, decompositionAddr, setAddr);
   sc_memory_arc_new(context, sc_type_arc_pos_const_perm, noroleAddr, decompositionEdge);
   sc_memory_arc_new(context, sc_type_arc_pos_const_perm, keynode_quasybinary_relation, noroleAddr);
   sc_memory_arc_new(context, sc_type_arc_pos_const_perm, keynode_nonbinary_relation, noroleAddr);
@@ -488,11 +484,7 @@ TEST_F(ScMemoryTest, agent_search_links_of_relation_connected_with_element)
   EXPECT_TRUE(sc_iterator5_next(it5));
   sc_iterator5_free(it5);
 
-  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(
-      context,
-      sc_iterator5_value(it5, 2),
-      sc_type_arc_pos_const_perm,
-      0);
+  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(context, sc_iterator5_value(it5, 2), sc_type_arc_pos_const_perm, 0);
 
   EXPECT_TRUE(sc_iterator3_next(it3));
   EXPECT_TRUE(sc_iterator3_next(it3));
@@ -522,7 +514,10 @@ TEST_F(ScMemoryTest, agent_search_decomposition)
 {
   sc_memory_context * context = sc_memory_context_new(sc_access_lvl_make_min);
 
-  sc_module_initialize();
+  sc_addr init_memory_generated_structure;
+  SC_ADDR_MAKE_EMPTY(init_memory_generated_structure);
+
+  sc_module_initialize_with_init_memory_generated_structure(init_memory_generated_structure);
 
   sc_addr const question = sc_memory_node_new(context, sc_type_node | sc_type_const);
   sc_addr const setAddr = sc_memory_node_new(context, sc_type_node | sc_type_const);
@@ -558,11 +553,7 @@ TEST_F(ScMemoryTest, agent_search_decomposition)
   EXPECT_TRUE(sc_iterator5_next(it5));
   sc_iterator5_free(it5);
 
-  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(
-      context,
-      sc_iterator5_value(it5, 2),
-      sc_type_arc_pos_const_perm,
-      0);
+  sc_iterator3 * it3 = sc_iterator3_f_a_a_new(context, sc_iterator5_value(it5, 2), sc_type_arc_pos_const_perm, 0);
 
   EXPECT_TRUE(sc_iterator3_next(it3));
   EXPECT_TRUE(sc_iterator3_next(it3));