+++ sc-memory/sc-core/sc_memory.c
@@ -23,6 +23,8 @@
 #include "sc-store/sc-base/sc_assert_utils.h"
 #include "sc-store/sc-base/sc_message.h"
 
+#include "sc_helper.h"
+
 sc_memory_context * s_memory_default_ctx = null_ptr;
 sc_uint16 s_context_id_last = 1;
 sc_uint32 s_context_id_count = 0;
@@ -73,14 +75,6 @@ sc_memory_context * sc_memory_initialize(const sc_memory_params * params)
     goto error;
   }
 
-  if (params->ext_path)
-  {
-    if (sc_memory_init_ext(params->ext_path, params->enabled_exts) == SC_RESULT_OK)
-    {
-      return s_memory_default_ctx;
-    }
-  }
-
   return s_memory_default_ctx;
 
 error:
@@ -92,10 +86,13 @@ error:
   return (s_memory_default_ctx = null_ptr);
 }
 
-sc_result sc_memory_init_ext(sc_char const * ext_path, const sc_char ** enabled_list)
+sc_result sc_memory_init_ext(
+    sc_char const * ext_path,
+    sc_char const ** enabled_list,
+    sc_addr const init_memory_generated_structure)
 {
   sc_result ext_res;
-  ext_res = sc_ext_initialize(ext_path, enabled_list);
+  ext_res = sc_ext_initialize(ext_path, enabled_list, init_memory_generated_structure);
 
   switch (ext_res)
   {