+++ sc-memory/sc-core/sc_memory_ext.c
@@ -23,6 +23,7 @@ typedef struct _sc_module_info
   gchar * path;
   sc_uint32 priority;
   fModuleFunc init_func;
+  char * init_func_type;
   fModuleFunc shut_func;
 } sc_module_info;
 
@@ -55,7 +56,10 @@ gint sc_priority_great(gconstpointer a, gconstpointer b)
   return sc_priority_less(b, a);
 }
 
-sc_result sc_ext_initialize(const sc_char * ext_dir_path, const sc_char ** enabled_list)
+sc_result sc_ext_initialize(
+    sc_char const * ext_dir_path,
+    sc_char const ** enabled_list,
+    sc_addr const init_memory_generated_structure)
 {
   GDir * ext_dir = null_ptr;
   const sc_char * file_name = null_ptr;
@@ -127,10 +131,19 @@ sc_result sc_ext_initialize(const sc_char * ext_dir_path, const sc_char ** enabl
     // skip non module files
     if (g_str_has_suffix(file_name, moduleSuffix) == SC_TRUE)
     {
-      if (g_module_symbol(mi->ptr, "sc_module_initialize", (gpointer *)&func) == SC_FALSE)
+      if (g_module_symbol(mi->ptr, "sc_module_initialize_with_init_memory_generated_structure", (gpointer *)&func) ==
+          SC_TRUE)
       {
-        sc_warning("Can't find 'sc_module_initialize' symbol in module: %s", mi->path);
-        goto clean;
+        mi->init_func_type = "sc_module_initialize_with_init_memory_generated_structure";
+      }
+      else
+      {
+        if (g_module_symbol(mi->ptr, "sc_module_initialize", (gpointer *)&func) == SC_FALSE)
+        {
+          sc_warning("Can't find any 'sc_module_initialize' symbol in module: %s", mi->path);
+          goto clean;
+        }
+        mi->init_func_type = "sc_module_initialize";
       }
       mi->init_func = func;
 
@@ -166,11 +179,21 @@ sc_result sc_ext_initialize(const sc_char * ext_dir_path, const sc_char ** enabl
 
   // initialize modules
   GList * item = modules_priority_list;
+  sc_result init_result = SC_RESULT_ERROR;
   while (item != null_ptr)
   {
     sc_module_info * module = (sc_module_info *)item->data;
     sc_message("Initialize module: %s", module->path);
-    if (module->init_func() != SC_RESULT_OK)
+    if (strcmp(module->init_func_type, "sc_module_initialize_with_init_memory_generated_structure") == 0)
+    {
+      init_result = module->init_func(init_memory_generated_structure);
+    }
+    else
+    {
+      init_result = module->init_func();
+    }
+
+    if (init_result != SC_RESULT_OK)
     {
       sc_warning("Something happens, on module initialization: %s", module->path);
       module->shut_func();