+++ sc-memory/sc-memory/sc_keynodes.cpp
@@ -63,21 +63,17 @@ bool ScKeynodes::Init(bool force, sc_char const * init_memory_generated_structur
     initMemoryGeneratedStructure =
         ctx.HelperResolveSystemIdtf(init_memory_generated_structure, ScType::NodeConstStruct);
     initMemoryGeneratedStructureValid = initMemoryGeneratedStructure.IsValid();
+    if (initMemoryGeneratedStructureValid)
+    {
+      ScAddr kNrelSysIdtf = ctx.HelperResolveSystemIdtf("nrel_system_identifier", ScType::NodeConstNoRole);
+      if (!ctx.HelperCheckEdge(initMemoryGeneratedStructure, kNrelSysIdtf, ScType::EdgeAccessConstPosPerm))
+      {
+        ctx.CreateEdge(ScType::EdgeAccessConstPosPerm, initMemoryGeneratedStructure, kNrelSysIdtf);
+      }
+    }
   }
 
   bool result = ScKeynodes::InitGlobal(initMemoryGeneratedStructure);
-
-  ScAddrVector const & keynodesAddrs = {
-      kCommandStateAddr,
-      kCommandInitiatedAddr,
-      kCommandProgressedAddr,
-      kCommandFinishedAddr,
-      kNrelResult,
-      kNrelCommonTemplate,
-      kNrelIdtf,
-      kNrelFormat,
-      kScResult,
-      kBinaryType};
 
   ScAddrVector const & resultCodes = {
       kScResultError,