+++ sc-tools/sc-config-utils/sc-config/sc_config.cpp
@@ -7,10 +7,9 @@
 #include "sc_config.hpp"
 
 #include <string>
-#include <sstream>
 #include <utility>
 #include <algorithm>
-#include <vector>
+#include <unordered_set>
 
 extern "C"
 {
@@ -21,11 +20,13 @@ extern "C"
 ScConfigGroup::ScConfigGroup(
     sc_config * config,
     std::string configPath,
-    std::vector<std::string> pathKeys,
+    std::unordered_set<std::string> const & pathKeys,
+    std::unordered_set<std::string> const & notUsedKeys,
     std::string group)
   : m_config(config)
   , m_configPath(std::move(configPath))
-  , m_pathKeys(std::move(pathKeys))
+  , m_pathKeys(pathKeys)
+  , m_notUsedKeys(notUsedKeys)
   , m_group(std::move(group))
 {
   sc_list * keys = sc_config_get_keys(m_config);
@@ -36,7 +37,10 @@ ScConfigGroup::ScConfigGroup(
     if (sc_str_has_prefix((sc_char *)element->data, m_group.c_str()))
     {
       std::string groupKey = std::string((sc_char *)element->data);
-      m_keys.push_back(groupKey.erase(0, m_group.size() + 1));
+
+      std::string const & key = groupKey.erase(0, m_group.size() + 1);
+      if (m_notUsedKeys.find(key) == m_notUsedKeys.cend())
+        m_keys.insert(key);
     }
 
     element = element->next;
@@ -50,7 +54,7 @@ std::string ScConfigGroup::operator[](std::string const & key) const
   std::string const & value = sc_config_get_value_string(m_config, m_group.c_str(), key.c_str());
   std::stringstream stream;
 
-  auto const & it = std::find(m_pathKeys.cbegin(), m_pathKeys.cend(), key);
+  auto const & it = m_pathKeys.find(key);
 
   if (it == m_pathKeys.cend())
     return value;
@@ -63,14 +67,18 @@ std::string ScConfigGroup::operator[](std::string const & key) const
   return stream.str();
 }
 
-std::vector<std::string> ScConfigGroup::operator*() const
+std::unordered_set<std::string> ScConfigGroup::operator*() const
 {
   return m_keys;
 }
 
-ScConfig::ScConfig(std::string path, std::vector<std::string> paths)
+ScConfig::ScConfig(
+    std::string path,
+    std::unordered_set<std::string> const & pathKeys,
+    std::unordered_set<std::string> const & notUsedKeys)
   : m_path(std::move(path))
-  , m_pathKeys(std::move(paths))
+  , m_pathKeys(pathKeys)
+  , m_notUsedKeys(notUsedKeys)
 {
   m_result = sc_config_initialize(&m_instance, m_path.c_str());
 }
@@ -82,7 +90,7 @@ sc_bool ScConfig::IsValid() const
 
 ScConfigGroup ScConfig::operator[](std::string const & group) const
 {
-  return ScConfigGroup(this->m_instance, this->GetDirectory(), this->m_pathKeys, group);
+  return ScConfigGroup(this->m_instance, this->GetDirectory(), this->m_pathKeys, this->m_notUsedKeys, group);
 }
 
 ScConfig::~ScConfig()