+++ sc-tools/sc-config-utils/sc_builder_config.hpp
@@ -14,9 +14,7 @@
 #include "sc-config/sc_config.hpp"
 
 #include "sc-core/sc_memory_params.h"
-
-#define DEFAULT_RESULT_STRUCTURE_MAIN_IDTF ""
-#define DEFAULT_RESULT_STRUCTURE_UPLOAD SC_FALSE
+#include "../sc-builder/src/builder.hpp"
 
 class ScBuilderConfig
 {
@@ -40,14 +38,6 @@ public:
 
   BuilderParams GetParams()
   {
-    m_builderParams.m_resultStructureUpload =
-        m_params.count("result_structure_upload")
-            ? (m_params.at("result_structure_upload") == "true" ? SC_TRUE : SC_FALSE)
-            : DEFAULT_RESULT_STRUCTURE_UPLOAD;
-    m_builderParams.m_resultStructureSystemIdtf = m_params.count("result_structure_system_idtf")
-                                                      ? m_params.at("result_structure_system_idtf").c_str()
-                                                      : DEFAULT_RESULT_STRUCTURE_MAIN_IDTF;
-
     return m_builderParams;
   }